refactor: split the stats module into format, plugin, and pruning files.
The tantivy plugin that writes and merges the component now lives apart from what the planner and executor read from it.
