+++ pg_search/src/index/stats/tests.rs
@@ -14,724 +14,21 @@
 //
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-//! Per-segment statistics: the `.stats` component.
-//!
-//! One `CompositeFile` per segment, keyed by `(Field, idx)`. `idx = 0` holds the empirical
-//! `min`/`max` of a fast field, `idx = 1` the logical box a partitioned build assigned to the
-//! segment's partition, and `idx = 2` stays reserved for sketches. The footer maps each entry to a
-//! byte range, so a reader touches only the entries it asks for. A missing file or entry means
-//! "unknown", and a consumer must keep the segment.
-//!
-//! The two entries have different lifecycles. Empirical stats come from the segment's own
-//! `.fast` file, so every immutable segment gets them, at write and at merge. Logical bounds come
-//! from the build that routed the rows; a merge keeps them only when every source has them,
-//! widened to the union box, which still holds every row.
-
-use std::any::Any;
-use std::cmp::Ordering;
-use std::collections::BTreeMap;
-use std::io::{self, Write};
-use std::ops::Bound;
-use std::sync::Arc;
-
-use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
-use tantivy::columnar::{Cardinality, ColumnarReader, DynamicColumn, DynamicColumnHandle};
-use tantivy::directory::error::OpenReadError;
-use tantivy::directory::{CompositeFile, CompositeWrite, FileSlice};
-use tantivy::index::{Segment, SegmentComponent, SegmentId, SegmentReader};
-use tantivy::indexer::DocIdMapping;
-use tantivy::schema::{Field, FieldType, Schema};
-use tantivy::{
-    Index, PluginMergeContext, PluginWriter, PluginWriterContext, SegmentPlugin, TantivyError,
-};
-
-use crate::api::HashMap;
-use crate::index::mvcc::MvccSatisfies;
-use crate::index::reader::index::SearchIndexReader;
-use crate::postgres::datetime::PostgresDateTime;
-use crate::postgres::pdb_owned_value::{PdbOwnedValue, exact_scalar_wire};
-use crate::postgres::rel::PgSearchRelation;
-use crate::postgres::storage::block::STATS_EXT;
-use crate::postgres::types::is_datetime_type;
-use crate::scan::range_partitioning::RangePartitioning;
-use crate::schema::SearchFieldType;
-
-const EMPIRICAL_IDX: usize = 0;
-const LOGICAL_IDX: usize = 1;
-
-/// Empirical `min`/`max` of one fast field within one segment. `nullable` is false only when
-/// every document holds exactly one value, so a segment without it may hide NULLs.
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct EmpiricalStats {
-    pub(crate) min: PdbOwnedValue,
-    pub(crate) max: PdbOwnedValue,
-    pub(crate) nullable: bool,
-}
-
-/// The half-open box a partitioned build assigned to a segment's partition on one field.
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct LogicalBounds {
-    pub(crate) lower: Bound<PdbOwnedValue>,
-    pub(crate) upper: Bound<PdbOwnedValue>,
-}
-
-/// Logical bounds keyed by field name, in the order the file lays them out.
-pub(crate) type LogicalBoundsByField = BTreeMap<String, LogicalBounds>;
-
-/// The default `PdbOwnedValue` serde is lossy (a non-negative `I64` comes back as `U64`), so the
-/// entries carry the exact scalar wire form the kd-tree also ships over the DSM.
-#[derive(Serialize, Deserialize)]
-struct WireScalar(#[serde(with = "exact_scalar_wire")] PdbOwnedValue);
-
-#[derive(Serialize, Deserialize)]
-struct EmpiricalWire {
-    nullable: bool,
-    min: WireScalar,
-    max: WireScalar,
-}
-
-#[derive(Serialize, Deserialize)]
-struct LogicalWire {
-    lower: Bound<WireScalar>,
-    upper: Bound<WireScalar>,
-}
-
-impl From<EmpiricalStats> for EmpiricalWire {
-    fn from(stats: EmpiricalStats) -> Self {
-        Self {
-            nullable: stats.nullable,
-            min: WireScalar(stats.min),
-            max: WireScalar(stats.max),
-        }
-    }
-}
-
-impl From<EmpiricalWire> for EmpiricalStats {
-    fn from(wire: EmpiricalWire) -> Self {
-        Self {
-            min: wire.min.0,
-            max: wire.max.0,
-            nullable: wire.nullable,
-        }
-    }
-}
-
-impl From<&LogicalBounds> for LogicalWire {
-    fn from(bounds: &LogicalBounds) -> Self {
-        Self {
-            lower: bounds.lower.as_ref().map(|v| WireScalar(v.clone())),
-            upper: bounds.upper.as_ref().map(|v| WireScalar(v.clone())),
-        }
-    }
-}
-
-impl From<LogicalWire> for LogicalBounds {
-    fn from(wire: LogicalWire) -> Self {
-        Self {
-            lower: wire.lower.map(|v| v.0),
-            upper: wire.upper.map(|v| v.0),
-        }
-    }
-}
-
-/// True when `hi` ends before `lo` starts, so two ranges with these ends cannot share a value.
-fn ends_before(hi: Bound<&PdbOwnedValue>, lo: Bound<&PdbOwnedValue>) -> bool {
-    match (hi, lo) {
-        (Bound::Unbounded, _) | (_, Bound::Unbounded) => false,
-        (Bound::Included(h), Bound::Included(l)) => h.total_cmp(l) == Ordering::Less,
-        (Bound::Included(h), Bound::Excluded(l))
-        | (Bound::Excluded(h), Bound::Included(l))
-        | (Bound::Excluded(h), Bound::Excluded(l)) => h.total_cmp(l) != Ordering::Greater,
-    }
-}
-
-fn ranges_intersect(
-    a_lo: Bound<&PdbOwnedValue>,
-    a_hi: Bound<&PdbOwnedValue>,
-    b_lo: Bound<&PdbOwnedValue>,
-    b_hi: Bound<&PdbOwnedValue>,
-) -> bool {
-    !ends_before(a_hi, b_lo) && !ends_before(b_hi, a_lo)
-}
-
-/// The looser of two lower bounds. At an equal value the inclusive one is looser.
-fn min_lower(a: &Bound<PdbOwnedValue>, b: &Bound<PdbOwnedValue>) -> Bound<PdbOwnedValue> {
-    match (a, b) {
-        (Bound::Unbounded, _) | (_, Bound::Unbounded) => Bound::Unbounded,
-        (Bound::Included(x), Bound::Included(y)) | (Bound::Excluded(x), Bound::Excluded(y)) => {
-            if x.total_cmp(y) == Ordering::Greater {
-                b.clone()
-            } else {
-                a.clone()
-            }
-        }
-        (Bound::Included(x), Bound::Excluded(y)) => {
-            if x.total_cmp(y) == Ordering::Greater {
-                b.clone()
-            } else {
-                a.clone()
-            }
-        }
-        (Bound::Excluded(x), Bound::Included(y)) => {
-            if y.total_cmp(x) == Ordering::Greater {
-                a.clone()
-            } else {
-                b.clone()
-            }
-        }
-    }
-}
-
-/// The looser of two upper bounds. At an equal value the inclusive one is looser.
-fn max_upper(a: &Bound<PdbOwnedValue>, b: &Bound<PdbOwnedValue>) -> Bound<PdbOwnedValue> {
-    match (a, b) {
-        (Bound::Unbounded, _) | (_, Bound::Unbounded) => Bound::Unbounded,
-        (Bound::Included(x), Bound::Included(y)) | (Bound::Excluded(x), Bound::Excluded(y)) => {
-            if x.total_cmp(y) == Ordering::Less {
-                b.clone()
-            } else {
-                a.clone()
-            }
-        }
-        (Bound::Included(x), Bound::Excluded(y)) => {
-            if x.total_cmp(y) == Ordering::Less {
-                b.clone()
-            } else {
-                a.clone()
-            }
-        }
-        (Bound::Excluded(x), Bound::Included(y)) => {
-            if y.total_cmp(x) == Ordering::Less {
-                a.clone()
-            } else {
-                b.clone()
-            }
-        }
-    }
-}
-
-/// Whether `total_cmp` ranks these two values by value. The derived order it falls back to
-/// ranks by variant, so a bound of one kind against a statistic of another says nothing.
-fn comparable(a: &PdbOwnedValue, b: &PdbOwnedValue) -> bool {
-    use PdbOwnedValue::*;
-    matches!(
-        (a, b),
-        (I64(_) | U64(_), I64(_) | U64(_))
-            | (F64(_), F64(_))
-            | (Str(_), Str(_))
-            | (Bytes(_), Bytes(_))
-            | (Bool(_), Bool(_))
-            | (Date(_), Date(_))
-            | (IpAddr(_), IpAddr(_))
-    )
-}
-
-fn bound_comparable(bound: &Bound<PdbOwnedValue>, value: &PdbOwnedValue) -> bool {
-    match bound {
-        Bound::Unbounded => true,
-        Bound::Included(b) | Bound::Excluded(b) => comparable(b, value),
-    }
-}
-
-impl LogicalBounds {
-    /// The smallest box that holds both.
-    pub(crate) fn union(&self, other: &Self) -> Self {
-        Self {
-            lower: min_lower(&self.lower, &other.lower),
-            upper: max_upper(&self.upper, &other.upper),
-        }
-    }
-
-    /// Whether a value in this box can fall in `[lower, upper)`. Bounds of a kind this box
-    /// cannot be ranked against count as intersecting.
-    pub(crate) fn intersects(
-        &self,
-        lower: &Bound<PdbOwnedValue>,
-        upper: &Bound<PdbOwnedValue>,
-    ) -> bool {
-        for own in [&self.lower, &self.upper] {
-            if let Bound::Included(v) | Bound::Excluded(v) = own
-                && !(bound_comparable(lower, v) && bound_comparable(upper, v))
-            {
-                return true;
-            }
-        }
-        ranges_intersect(
-            self.lower.as_ref(),
-            self.upper.as_ref(),
-            lower.as_ref(),
-            upper.as_ref(),
-        )
-    }
-
-    /// A build routes NULLs below every split, so only a box open at the bottom can hold them.
-    pub(crate) fn may_hold_nulls(&self) -> bool {
-        matches!(self.lower, Bound::Unbounded)
-    }
-}
-
-impl EmpiricalStats {
-    /// Whether a value in `[min, max]` can fall in `[lower, upper)`. Bounds of a kind these
-    /// statistics cannot be ranked against count as intersecting.
-    pub(crate) fn intersects(
-        &self,
-        lower: &Bound<PdbOwnedValue>,
-        upper: &Bound<PdbOwnedValue>,
-    ) -> bool {
-        if !(bound_comparable(lower, &self.min) && bound_comparable(upper, &self.max)) {
-            return true;
-        }
-        ranges_intersect(
-            Bound::Included(&self.min),
-            Bound::Included(&self.max),
-            lower.as_ref(),
-            upper.as_ref(),
-        )
-    }
-
-    /// Datetimes of a recent index sit in an `I64` column; a query bound on that field is a
-    /// `Date`, so lift the statistics the way `FFType` lifts the column's values.
-    fn into_dates(self) -> Option<Self> {
-        let lift = |v: PdbOwnedValue| match v {
-            PdbOwnedValue::I64(raw) => PostgresDateTime::try_from_raw(raw)
-                .ok()
-                .map(PdbOwnedValue::Date),
-            other => Some(other),
-        };
-        Some(Self {
-            min: lift(self.min)?,
-            max: lift(self.max)?,
-            nullable: self.nullable,
-        })
-    }
-}
-
-/// The plugin that writes and merges the `.stats` component.
-pub(crate) struct StatsPlugin;
-
-/// Attaches the plugin to an index. Every `Index` pg_search writes or merges through must call
-/// this: tantivy restores only its built-in plugins, and the index metadata lists `stats` as
-/// required. A read-only `Index` has no use for it.
-pub(crate) fn register(index: &mut Index) {
-    index.register_plugin(Arc::new(StatsPlugin));
-}
-
-impl SegmentPlugin for StatsPlugin {
-    fn extensions(&self) -> &[&str] {
-        &[STATS_EXT]
-    }
-
-    fn create_writer(&self, _ctx: &PluginWriterContext) -> tantivy::Result<Box<dyn PluginWriter>> {
-        Ok(Box::new(StatsWriter { logical: None }))
-    }
-
-    fn merge(&self, ctx: PluginMergeContext) -> tantivy::Result<()> {
-        let logical = merged_logical_bounds(ctx.readers, ctx.schema)?;
-        write_stats(ctx.target_segment, ctx.schema, logical.as_ref())
-    }
-}
-
-/// Per-segment writer. It records nothing per document: the statistics come off the finished
-/// `.fast` file, which the built-in plugins serialize before this one runs.
-pub(crate) struct StatsWriter {
-    logical: Option<Arc<LogicalBoundsByField>>,
-}
-
-impl StatsWriter {
-    pub(crate) fn set_logical_bounds(&mut self, bounds: Arc<LogicalBoundsByField>) {
-        self.logical = Some(bounds);
-    }
-}
-
-impl PluginWriter for StatsWriter {
-    fn serialize(
-        self: Box<Self>,
-        segment: &Segment,
-        _doc_id_map: Option<&DocIdMapping>,
-    ) -> tantivy::Result<()> {
-        write_stats(segment, &segment.schema(), self.logical.as_deref())
-    }
-
-    fn mem_usage(&self) -> usize {
-        0
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn as_any_mut(&mut self) -> &mut dyn Any {
-        self
-    }
-}
-
-fn stats_component() -> SegmentComponent {
-    SegmentComponent::Custom(STATS_EXT.to_string())
-}
-
-fn encode<T: Serialize>(value: &T) -> tantivy::Result<Vec<u8>> {
-    postcard::to_allocvec(value).map_err(|e| TantivyError::InternalError(e.to_string()))
-}
-
-fn write_stats(
-    segment: &Segment,
-    schema: &Schema,
-    logical: Option<&LogicalBoundsByField>,
-) -> tantivy::Result<()> {
-    let empirical = empirical_stats(segment, schema)?;
-    let mut write = CompositeWrite::wrap(segment.open_write(stats_component())?);
-    for (field, stats) in empirical {
-        let bytes = encode(&EmpiricalWire::from(stats))?;
-        write
-            .for_field_with_idx(field, EMPIRICAL_IDX)
-            .write_all(&bytes)?;
-    }
-    for (name, bounds) in logical.into_iter().flatten() {
-        let Ok(field) = schema.get_field(name) else {
-            continue;
-        };
-        if !logical_bounds_hold(schema, field) {
-            continue;
-        }
-        let bytes = encode(&LogicalWire::from(bounds))?;
-        write
-            .for_field_with_idx(field, LOGICAL_IDX)
-            .write_all(&bytes)?;
-    }
-    write.close()?;
-    Ok(())
-}
-
-/// The build routes on raw values, but a partition's range query runs on the fast column. A
-/// text normalizer other than `raw` reorders that column, so a box in raw order could prune a
-/// segment the query still needs.
-fn logical_bounds_hold(schema: &Schema, field: Field) -> bool {
-    match schema.get_field_entry(field).field_type() {
-        FieldType::Str(options) => options.get_fast_field_tokenizer_name() == Some("raw"),
-        _ => true,
-    }
-}
-
-/// `min`/`max` of every fast field column of `segment`, read off its `.fast` file.
-fn empirical_stats(
-    segment: &Segment,
-    schema: &Schema,
-) -> tantivy::Result<Vec<(Field, EmpiricalStats)>> {
-    let fast = match segment.open_read(SegmentComponent::FastFields) {
-        Ok(fast) => fast,
-        Err(OpenReadError::FileDoesNotExist(_)) => return Ok(Vec::new()),
-        Err(e) => return Err(e.into()),
-    };
-    let columnar = ColumnarReader::open(fast)?;
-    let mut by_field: HashMap<Field, Vec<DynamicColumnHandle>> = HashMap::default();
-    for (name, handle) in columnar.iter_columns()? {
-        // JSON subpaths carry no schema field of their own.
-        let Ok(field) = schema.get_field(&name) else {
-            continue;
-        };
-        by_field.entry(field).or_default().push(handle);
-    }
-    let mut stats = Vec::with_capacity(by_field.len());
-    for (field, handles) in by_field {
-        // A field spread over several typed columns has no single order to bound.
-        let [handle] = handles.as_slice() else {
-            continue;
-        };
-        if let Some(column_stats) = column_stats(handle)? {
-            stats.push((field, column_stats));
-        }
-    }
-    stats.sort_by_key(|(field, _)| field.field_id());
-    Ok(stats)
-}
-
-fn column_stats(handle: &DynamicColumnHandle) -> tantivy::Result<Option<EmpiricalStats>> {
-    let (min, max, cardinality) = match handle.open()? {
-        DynamicColumn::I64(c) => (
-            PdbOwnedValue::I64(c.min_value()),
-            PdbOwnedValue::I64(c.max_value()),
-            c.get_cardinality(),
-        ),
-        DynamicColumn::U64(c) => (
-            PdbOwnedValue::U64(c.min_value()),
-            PdbOwnedValue::U64(c.max_value()),
-            c.get_cardinality(),
-        ),
-        DynamicColumn::F64(c) => (
-            PdbOwnedValue::F64(c.min_value()),
-            PdbOwnedValue::F64(c.max_value()),
-            c.get_cardinality(),
-        ),
-        DynamicColumn::Bool(c) => (
-            PdbOwnedValue::Bool(c.min_value()),
-            PdbOwnedValue::Bool(c.max_value()),
-            c.get_cardinality(),
-        ),
-        DynamicColumn::DateTime(c) => {
-            // A value Postgres can't represent has no place in a segment, but a statistic is
-            // not worth failing the write over.
-            let (Ok(min), Ok(max)) = (
-                PostgresDateTime::try_from(c.min_value()),
-                PostgresDateTime::try_from(c.max_value()),
-            ) else {
-                return Ok(None);
-            };
-            (
-                PdbOwnedValue::Date(min),
-                PdbOwnedValue::Date(max),
-                c.get_cardinality(),
-            )
-        }
-        DynamicColumn::IpAddr(c) => (
-            PdbOwnedValue::IpAddr(c.min_value()),
-            PdbOwnedValue::IpAddr(c.max_value()),
-            c.get_cardinality(),
-        ),
-        DynamicColumn::Bytes(c) => {
-            let num_terms = c.num_terms();
-            if num_terms == 0 {
-                return Ok(None);
-            }
-            let (mut min, mut max) = (Vec::new(), Vec::new());
-            c.ord_to_bytes(0, &mut min)?;
-            c.ord_to_bytes(num_terms as u64 - 1, &mut max)?;
-            (
-                PdbOwnedValue::Bytes(min),
-                PdbOwnedValue::Bytes(max),
-                c.ords().get_cardinality(),
-            )
-        }
-        DynamicColumn::Str(c) => {
-            let num_terms = c.num_terms();
-            if num_terms == 0 {
-                return Ok(None);
-            }
-            let (mut min, mut max) = (String::new(), String::new());
-            c.ord_to_str(0, &mut min)?;
-            c.ord_to_str(num_terms as u64 - 1, &mut max)?;
-            (
-                PdbOwnedValue::Str(min),
-                PdbOwnedValue::Str(max),
-                c.ords().get_cardinality(),
-            )
-        }
-    };
-    Ok(Some(EmpiricalStats {
-        min,
-        max,
-        nullable: cardinality != Cardinality::Full,
-    }))
-}
-
-/// The union of the sources' logical bounds, per field, or `None` when any source has no
-/// `.stats` at all. A field is kept only when every source bounds it.
-fn merged_logical_bounds(
-    readers: &[SegmentReader],
-    schema: &Schema,
-) -> tantivy::Result<Option<LogicalBoundsByField>> {
-    let mut sources = Vec::with_capacity(readers.len());
-    for reader in readers {
-        match SegmentStats::of_reader(reader)? {
-            Some(stats) => sources.push(stats),
-            // A segment written before the component existed. The merge cannot claim a box it
-            // cannot prove for it.
-            None => return Ok(None),
-        }
-    }
-    if sources.is_empty() {
-        return Ok(None);
-    }
-    let mut merged = LogicalBoundsByField::new();
-    'fields: for (field, entry) in schema.fields() {
-        let mut union: Option<LogicalBounds> = None;
-        for source in &sources {
-            let Some(bounds) = source.logical(field)? else {
-                continue 'fields;
-            };
-            union = Some(match union {
-                None => bounds,
-                Some(current) => current.union(&bounds),
-            });
-        }
-        if let Some(union) = union {
-            merged.insert(entry.name().to_string(), union);
-        }
-    }
-    Ok(Some(merged))
-}
-
-/// A segment's `.stats` file, opened on its footer. Entries are decoded on request.
-pub(crate) struct SegmentStats {
-    file: CompositeFile,
-}
-
-impl SegmentStats {
-    pub(crate) fn open(slice: FileSlice) -> io::Result<Self> {
-        Ok(Self {
-            file: CompositeFile::open(&slice)?,
-        })
-    }
-
-    /// The `.stats` of `segment`, through its directory. `None` when the segment was written
-    /// without the component.
-    pub(crate) fn of_segment(segment: &Segment) -> io::Result<Option<Self>> {
-        Self::from_component(segment.open_read(stats_component()))
-    }
-
-    /// The `.stats` of an open segment reader. `None` when the segment was written without the
-    /// component.
-    pub(crate) fn of_reader(reader: &SegmentReader) -> io::Result<Option<Self>> {
-        Self::from_component(reader.open_read(stats_component()))
-    }
-
-    fn from_component(opened: Result<FileSlice, OpenReadError>) -> io::Result<Option<Self>> {
-        match opened {
-            Ok(slice) => Self::open(slice).map(Some),
-            Err(OpenReadError::FileDoesNotExist(_)) => Ok(None),
-            Err(e) => Err(io::Error::other(e)),
-        }
-    }
-
-    pub(crate) fn empirical(&self, field: Field) -> io::Result<Option<EmpiricalStats>> {
-        Ok(self
-            .read::<EmpiricalWire>(field, EMPIRICAL_IDX)?
-            .map(EmpiricalStats::from))
-    }
-
-    pub(crate) fn logical(&self, field: Field) -> io::Result<Option<LogicalBounds>> {
-        Ok(self
-            .read::<LogicalWire>(field, LOGICAL_IDX)?
-            .map(LogicalBounds::from))
-    }
-
-    fn read<T: DeserializeOwned>(&self, field: Field, idx: usize) -> io::Result<Option<T>> {
-        let Some(slice) = self.file.open_read_with_idx(field, idx) else {
-            return Ok(None);
-        };
-        let bytes = slice.read_bytes()?;
-        postcard::from_bytes(&bytes)
-            .map(Some)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-    }
-}
-
-/// The split points a partitioned build stamped on the index: the edges of every box a visible
-/// segment carries for `partition_by`. `None` when no segment carries one. A segment without a
-/// box does not veto the grid, since its own statistics place it at execution.
-pub(crate) fn persisted_split_points(
-    indexrel: &PgSearchRelation,
-    partition_by: &str,
-) -> anyhow::Result<Option<Vec<PdbOwnedValue>>> {
-    if indexrel.options().partition_by().is_empty() {
-        return Ok(None);
-    }
-    let directory = MvccSatisfies::Snapshot.directory(indexrel);
-    let index = Index::open(directory.clone())?;
-    let Ok(field) = index.schema().get_field(partition_by) else {
-        return Ok(None);
-    };
-    let mut points = Vec::new();
-    for segment in index.searchable_segments()? {
-        // Opening a component of a mutable segment materializes the whole segment first, and its
-        // entry already says it has no `.stats`.
-        let has_stats = directory
-            .segment_meta_entry(&segment.id())
-            .is_some_and(|entry| entry.stats().is_some());
-        if !has_stats {
-            continue;
-        }
-        let Some(stats) = SegmentStats::of_segment(&segment)? else {
-            continue;
-        };
-        let Some(bounds) = stats.logical(field)? else {
-            continue;
-        };
-        for bound in [bounds.lower, bounds.upper] {
-            if let Bound::Included(v) | Bound::Excluded(v) = bound {
-                points.push(v);
-            }
-        }
-    }
-    points.sort_unstable_by(PdbOwnedValue::total_cmp);
-    points.dedup_by(|a, b| a.total_cmp(b) == Ordering::Equal);
-    Ok((!points.is_empty()).then_some(points))
-}
-
-/// The segments of `reader` that can hold a row of `partition`. A segment without statistics
-/// it can be ranked against is kept, so the query the caller still applies stays the source of
-/// truth.
-pub(crate) fn segments_for_partition(
-    reader: &SearchIndexReader,
-    boundaries: &RangePartitioning,
-    partition: usize,
-) -> Vec<SegmentId> {
-    let all = reader.segment_ids();
-    let Some((lower, upper)) = boundaries.partition_range(partition) else {
-        return all;
-    };
-    let field_name = boundaries.partition_by.as_ref();
-    let Ok(field) = reader.schema().tantivy_schema().get_field(field_name) else {
-        return all;
-    };
-    // A recent index keeps datetimes in an `I64` column, so its statistics need the same lift
-    // as its values; a legacy index stored them as `Date` already.
-    let is_date = match reader
-        .schema()
-        .search_field(field_name)
-        .map(|f| f.field_type())
-    {
-        Some(SearchFieldType::Date(_)) => true,
-        Some(SearchFieldType::I64(oid)) => is_datetime_type(oid),
-        _ => false,
-    };
-    // NULLs route to partition 0, so it keeps every segment that may hold one.
-    let catches_nulls = partition == 0;
-
-    reader
-        .segment_readers()
-        .iter()
-        .filter(|segment_reader| {
-            let Ok(Some(stats)) = SegmentStats::of_reader(segment_reader) else {
-                return true;
-            };
-            match stats.logical(field) {
-                Ok(Some(bounds)) => {
-                    return (catches_nulls && bounds.may_hold_nulls())
-                        || bounds.intersects(&lower, &upper);
-                }
-                Ok(None) => {}
-                Err(_) => return true,
-            }
-            let Ok(Some(empirical)) = stats.empirical(field) else {
-                return true;
-            };
-            let empirical = if is_date {
-                match empirical.into_dates() {
-                    Some(empirical) => empirical,
-                    None => return true,
-                }
-            } else {
-                empirical
-            };
-            (catches_nulls && empirical.nullable) || empirical.intersects(&lower, &upper)
-        })
-        .map(SegmentReader::segment_id)
-        .collect()
-}
-
-#[cfg(any(test, feature = "pg_test"))]
 #[pgrx::pg_schema]
 mod tests {
     use std::ops::Bound;
 
     use pgrx::prelude::*;
     use tantivy::Index;
+    use tantivy::index::SegmentId;
 
-    use super::*;
+    use super::super::*;
     use crate::api::FieldName;
+    use crate::index::mvcc::MvccSatisfies;
+    use crate::index::reader::index::SearchIndexReader;
+    use crate::postgres::rel::PgSearchRelation;
     use crate::query::SearchQueryInput;
+    use crate::scan::range_partitioning::RangePartitioning;
 
     fn open_index(index: &str) -> PgSearchRelation {
         let oid = Spi::get_one::<pg_sys::Oid>(&format!("SELECT '{index}'::regclass::oid"))
